Added tests for arrayComparision.. and refactored code:
Modified the operation function to accept a comparator function as a parameter, which makes the function more flexible and easier to understand.
Removed the isUnion parameter and replaced it with a comparator function, simplifying the logic and making the code more readable.
Updated the inBoth, inFirstOnly, and inSecondOnly functions to use the new operation function with their respective comparators.